Removed redundant mutexes from the observer example.
Further improvements could be made - currently the whole object
that has EventProducer has to be wrapped in a mutex to share it
between threads. So if the object has reentrant functions that
wouldn't need a mutex, it's still there anyway. It would be
ideal if there was a way to wrap the member EventProducer
in a mutex, and leave other APIs and the parent object lock
free.